commit before merge

+++ src/main/java/org/wtm/web/ticket/model/TicketHistoryUsage.java
@@ -28,5 +28,5 @@ public class TicketHistoryUsage extends BaseTimeEntity {
     private Ticket ticket;
 
     @Column(nullable = false)
-    private Integer amount;
+    private Long amount;
 }
